Have wiringPiISR and wiringPiISR2 return error on failure in initialization
The each function initializes the event listener loop in a spawned
listener thread. The new thread does not report errors back to the
original caller. So the function misses initialization errors.
The code is fixed to initialize the event listener in the original
thread before spawning the listener thread so that the function
can pick up errors and return them to the user.